# TINKERPOP-2682 Enable WS compression in .NET