@spmallette 2022-02-03
> WebSocket compression can be problematic due to attacks like CRIME/BREACH....I guess we should discuss how we want to deal with this in general as it's not specific to .NET. So, should we simply add similar notes to the Java/Python docs?

I wasn't aware of this issue. Maybe you should create an issue for this in JIRA so that it's not forgotten? I don't think compression is configurable in java at this point. cc/ @divijvaidya  

Thanks for doing some benchmarks. GraphSON network traffic size differences always look so crazy to me whenever we flip compression on. 😄 

VOTE +1