@FlorianHockmann 2022-02-04
> I wasn't aware of this issue. Maybe you should create an issue for this in JIRA so that it's not forgotten?

Done: [TINKERPOP-2700](https://issues.apache.org/jira/browse/TINKERPOP-2700)